convert serializer type information to internal files
