+++ src/jsonapi-serializer.skel.d.ts
@@ -1,0 +1,4 @@
+declare module 'jsonapi-serializer' {
+  const content: any;
+  export = content;
+}
