+++ src/mapper.ts
@@ -6,3 +6,7 @@ export {
   // Add future adapters here
   // Sequelize
 }
+
+// Export TS types for TS consumers
+export * from './interfaces';
+export * from './jsonapi-serializer';
